chore(new-ui): refactor after new TransactionQueue types
