+++ packages/new-polymath-ui/src/components/ModalConfirm/ModalConfirm.mdx
@@ -20,7 +20,7 @@ import { Demo } from './Demo';
     transactionQueue={{
       id: '111',
       name: 'Dividend Configuration',
-      status: types.HigherLevelTransactionStatus.Idle,
+      status: types.TransactionQueueStatus.Idle,
       transactions: [
         {
           id: '0',
