+++ packages/new-polymath-ui/src/components/ModalTransactionQueue/docs/ModalTransactionQueue.mdx
@@ -5,7 +5,7 @@ menu: Modals
 
 import { Playground, PropsTable } from 'docz';
 
-import { ModalTransactionQueue } from './';
+import { ModalTransactionQueue } from '../ModalTransactionQueue';
 import { Demo } from './Demo';
 
 # ModalTransactionQueue
