+++ packages/new-polymath-ui/src/docz/pages/TransactionQueue/TransactionQueueDemo.mdx
@@ -15,7 +15,7 @@ import { Demo } from './Demo';
     transactionQueue={{
       id: '111',
       name: 'Dividend Configuration',
-      status: types.HigherLevelTransactionStatus.Idle,
+      status: types.TransactionQueueStatus.Idle,
       transactions: [
         {
           id: '0',
